+++ src/main/resources/templates/admin/plantingSite.html
@@ -310,6 +310,13 @@
             </label>
             <br/>
             <label>
+                Edit Behavior
+                <select name="editVersion">
+                    <option value="1" selected>Original (some edits not allowed)</option>
+                </select>
+            </label>
+            <br/>
+            <label>
                 Subzone IDs to mark incomplete (comma-separated)
                 <input type="text" name="subzoneIdsToMarkIncomplete"/>
             </label>
